Restructure the use of types to adhere to the new convention
