+++ src/state/actions/daoCreator.ts
@@ -1,5 +1,6 @@
 import { newNotificationError } from "./notifications"
 import * as Arc from "../../lib/integrations/daoStack/arc"
+import { Founder } from "../../lib/integrations/daoStack/arc"
 
 // Stepper
 export const STEP_NEXT = "STEP_NEXT"
