--- src/types/founder.d.ts
@@ -1,5 +1,0 @@
-type Founder = {
-  address: string
-  reputation: string
-  tokens: string
-}
